fix: use already existing regexes
